+++ .gitignore
@@ -13,6 +13,10 @@ __debug_bin
 # JetBrains
 .idea/
 
+# Go workspace file
+go.work
+go.work.sum
+
 # configuration file
 /config/*.json
 /config/*.toml
